refactor(menu): order와 table 패키지 합침
