+++ src/main/java/kitchenpos/order/domain/OrderTables.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.domain;
+package kitchenpos.order.domain;
 
 import kitchenpos.tablegroup.domain.TableGroup;
 import org.springframework.util.CollectionUtils;
