+++ src/main/java/kitchenpos/order/domain/Orders.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.domain;
+package kitchenpos.order.domain;
 
 import kitchenpos.order.domain.Order;
