+++ src/main/java/kitchenpos/tablegroup/dto/TableGroupRequest.java
@@ -1,6 +1,6 @@
 package kitchenpos.tablegroup.dto;
 
-import kitchenpos.table.dto.OrderTableRequest;
+import kitchenpos.order.dto.OrderTableRequest;
 import org.springframework.util.CollectionUtils;
 
 import java.util.List;
