+++ src/test/java/kitchenpos/order/domain/OrderTablesTest.java
@@ -1,8 +1,6 @@
 package kitchenpos.order.domain;
 
 
-import kitchenpos.table.domain.OrderTable;
-import kitchenpos.table.domain.OrderTables;
 import kitchenpos.tablegroup.domain.TableGroup;
 import org.assertj.core.api.ThrowableAssert;
 import org.junit.jupiter.api.DisplayName;
